le graph malades/morts en fct de poucentage de vaccination est possible. Il faudra l'ameliorrer par contre

+++ trunk/M2/V1/psia/tmes/tme2/fr.upmc.sar.psia.tme2/src/fr/upmc/sar/psia/tme2/Grippe.java
@@ -21,8 +21,8 @@ public class Grippe implements EDProtocol {
 	public static final int MIN_VOISIN    = 8;
 	public static final int MAX_VOISIN    = 15;
 
-	public static Integer nbMort   = 0;
-	public static Integer nbMalade = 0;
+	public static Double nbMort   = 0.0;
+	public static Double nbMalade = 0.0;
 
 
 	// Identifiant de la couche transport
